Deprecated tmap functions

Since version 2.0, tmap function names are prefixed with a tm_ or tmap_. Therefore, function names used by tmap 1.x such as animation_tmap have been renamed to tmap_animation.

Details

animation_tmap: replaced by tmap_animation

save_tmap: replaced by tmap_save

style_catalogue: replaced by tmap_style_catalogue

style_catalog: replaced by tmap_style_catalog

last_map: replaced by tmap_last

tm_style_white: replaced by tm_style("white")

tm_style_gray: replaced by tm_style("gray")

tm_style_grey: replaced by tm_style("grey")

tm_style_natural: replaced by tm_style("natural")

tm_style_cobalt: replaced by tm_style("cobalt")

tm_style_col_blind: replaced by tm_style("col_blind")

tm_style_albatross: replaced by tm_style("albatross")

tm_style_beaver: replaced by tm_style("beaver")

tm_style_bw: replaced by tm_style("bw")

tm_style_classic: replaced by tm_style("classic")

tm_format_World: replaced by tm_format("World")

tm_format_World_wide: replaced by tm_format("World_wide")

tm_format_NLD: replaced by tm_format("NLD")

tm_format_NLD_wide: replaced by tm_format("NLD_wide")

tm_format_Europe: not used anymore, since the dataset Europe is no longer maintained

tm_format_Europe2: not used anymore, since the dataset Europe is no longer maintained

tm_format_Europe_wide: not used anymore, since the dataset Europe is no longer maintained