Map layer: cartogram

Map layer that draws a cartogram

Usage

tm_cartogram(
  size = 1,
  size.scale = tm_scale(),
  size.legend = tm_legend_hide(),
  size.free = NA,
  plot.order = tm_plot_order("size", reverse = FALSE),
  trans.args = list(type = "cont", itermax = 15),
  ...
)

Arguments

size, size.scale, size.legend, size.free

Transformation variable that determines the size of the polygons.

plot.order

Specification in which order the spatial features are drawn. See tm_plot_order() for details.

trans.args

lists that are passed on to internal transformation function

...

Arguments passed on to tm_polygons

fill,fill.scale,fill.legend,fill.free

Visual variable that determines the fill color. See details.

col,col.scale,col.legend,col.free

Visual variable that determines the border color. See details.

lwd,lwd.scale,lwd.legend,lwd.free

Visual variable that determines the line width. See details.

lty,lty.scale,lty.legend,lty.free

Visual variable that determines the line type. See details.

fill_alpha,fill_alpha.scale,fill_alpha.legend,fill_alpha.free

Visual variable that determines the fill color alpha transparency See details.

col_alpha,col_alpha.scale,col_alpha.legend,col_alpha.free

Visual variable that determines the border color alpha transparency. See details.

linejoin,lineend

Line join and line end. See gpar() for details.

trans.args,mapping.args

lists that are passed on to internal transformation and mapping functions respectively

zindex

Map layers are drawn on top of each other. The zindex numbers (one for each map layer) determines the stacking order. By default the map layers are drawn in the order they are called.

group

Name of the group to which this layer belongs. This is only relevant in view mode, where layer groups can be switched (see group.control)

group.control

In view mode, the group control determines how layer groups can be switched on and off. Options: "radio" for radio buttons (meaning only one group can be shown), "check" for check boxes (so multiple groups can be shown), and "none" for no control (the group cannot be (de)selected).

popup.vars

names of data variables that are shown in the popups in "view" mode. Set popup.vars to TRUE to show all variables in the shape object. Set popup.vars to FALSE to disable popups. Set popup.vars to a character vector of variable names to those those variables in the popups. The default (NA) depends on whether visual variables (e.g.col) are used. If so, only those are shown. If not all variables in the shape object are shown.

popup.format

list of formatting options for the popup values. See the argument legend.format for options. Only applicable for numeric data variables. If one list of formatting options is provided, it is applied to all numeric variables of popup.vars. Also, a (named) list of lists can be provided. In that case, each list of formatting options is applied to the named variable.

hover

name of the data variable that specifies the hover labels

id

name of the data variable that specifies the indices of the spatial features. Only used for "view" mode.