Map component: scale bar

This function was renamed to tm_scalebar() in tmap v4.0.

Usage

tm_scale_bar(...)

Arguments

...

Arguments passed on to tm_scalebar

breaks

breaks

width

width of the scale bar. Units are number of text line heights, which is similar to the number of characters.

text.size

text size

text.color

text.color

color.dark

color.dark

color.light

color.light

lwd

linewidth

size

Deprecated (use text.size instead)

stack

stack

margins

margins

frame

frame

frame.lwd

frame line width

frame.r

Radius of the rounded frame corners. 0 means no rounding.

bg.color

Background color

bg.alpha

Background transparency

position

Vector of two values, specifying the x and y coordinates. The first is either "left" or "right", the second either "top" or "bottom".

z

z