tmap_last {tmap}	R Documentation

Retrieve the last map to be modified or created

Description

Retrieve the last map to be modified or created. Works in the same way as ggplot2's last_plot, although there is a difference: last_map returns the last call instead of the stacked tmap-elements.

Usage

tmap_last()

Value

call

See Also

tmap_save

[Package tmap version 3.0 Index]