World {tmap}	R Documentation

World and Netherlands map

Description

Maps of the world and the Netherlands (province and municipality level), class sf

Usage

data(World)

data(NLD_prov)

data(NLD_muni)

Details

The default projections for these maps are Eckhart IV (World) and Rijksdriehoekstelsel (Netherlands). See below. The projection can be changed temporarily for plotting purposes by using the projection argument of tm_shape (or qtm).

World World map. The default projection for this world map is Eckhart IV since area sizes are preserved, which is a very important property for choropleths.

NLD_prov and NLD_muni, maps of the Netherlands at province and municipality level of 2013. The used projection is the Rijksdriehoekstelsel projection. Important: publication of these maps is only allowed when cited to Statistics Netherlands (CBS) and Kadaster Nederland as source.

Source

http://www.naturalearthdata.com for World

http://www.happyplanetindex.org for World

http://www.cbs.nl for NLD_prov and NLD_muni.

References

Statistics Netherlands (2014), The Hague/Heerlen, Netherlands, http://www.cbs.nl.

Kadaster, the Netherlands' Cadastre, Land Registry, and Mapping Agency (2014), Apeldoorn, Netherlands, http://www.kadaster.nl.

[Package tmap version 3.0 Index]