land {tmap}	R Documentation

Spatial data of global land cover

Description

Spatial data of global land cover, of class RasterBrick. The data includes a population times series from 1950 to (forecasted) 2030. All metro areas with over 1 million inhabitants in 2010 are included.

Usage

data(land)

Details

Important: publication of these maps is only allowed when cited to Tateishi et al. (2014), and when "Geospatial Information Authority of Japan, Chiba University and collaborating organizations." is shown. See http://www.iscgm.org/gm/glcnmo.html#use.

Source

http://www.iscgm.org/gm/glcnmo.html

References

Production of Global Land Cover Data - GLCNMO2008, Tateishi, R., Thanh Hoan, N., Kobayashi, T., Alsaaideh, B., Tana, G., Xuan Phong, D. (2014), Journal of Geography and Geology, 6 (3).

[Package tmap version 3.0 Index]