metro {tmap}	R Documentation

Spatial data of metropolitan areas

Description

Spatial data of metropolitan areas, of class sf. The data includes a population times series from 1950 to (forecasted) 2030. All metro areas with over 1 million inhabitants in 2010 are included.

Usage

data(metro)

Source

https://population.un.org/wup/

References

United Nations, Department of Economic and Social Affairs, Population Division (2014). World Urbanization Prospects: The 2014 Revision, CD-ROM Edition.

[Package tmap version 3.0 Index]