tm_add_legend {tmap}	R Documentation

Add manual legend

Description

Creates a tmap-element that adds a manual legend.

Usage

tm_add_legend(
  type = c("fill", "symbol", "text", "line"),
  labels = NULL,
  col = NULL,
  size = NULL,
  shape = NULL,
  lwd = NULL,
  lty = NULL,
  text = NULL,
  alpha = NA,
  border.col = "black",
  border.lwd = 1,
  border.alpha = NA,
  title = "",
  is.portrait = TRUE,
  legend.format = list(),
  reverse = FALSE,
  z = NA,
  group = NULL
)

Arguments

type
type of legend. One of "fill", "symbol", "text", or "line"

labels
legend labels

col
legend colors

size
legend symbol sizes (if type=="symbol")

shape
legend symbol shapes (if type=="symbol")

lwd
legend line widths (if type=="line")

lty
legend line types (if type=="line")

text
legend texts (if type=="text")

alpha
legend fill transparency

border.col
legend border col (if type is "fill" or "symbol")

border.lwd
legend border width (if type is "fill" or "symbol")

border.alpha
legend border alpha (if type is "fill" or "symbol")

title
legend title

is.portrait
is legend portrait (TRUE) or landscape (FALSE)?

legend.format
options to format the legend, see tm_symbols (the description of the argument legend.format) for details. Note that many of these arguments are not applicable for tm_add_legend since labels should be a character vector. However, some options could still be handy, e.g. list(text.align = "right").

reverse
are the legend items reversed (by default FALSE)?

z
legend stack position

group
name of the group to which this layer belongs in view mode. Each group can be selected or deselected in the layer control item. By default NULL, which means that the legend will not be shown in the layer control item.

See Also

tm_symbols for an example

[Package tmap version 3.0 Index]