Specify an animation (experimental)

tm_animate

Usage

tm_animate(
  by = "VARS__",
  nframes = 60L,
  fps = 20L,
  play = c("loop", "pingpong", "once"),
  dpr = 2,
  ...
)

tm_animate_slow(
  by = "VARS__",
  nframes = 60L,
  fps = 2L,
  play = c("loop", "pingpong", "once"),
  dpr = 2,
  ...
)

Arguments

by

group by variable used to create the animation frames. Note: it is called pages in the core function tm_facets().

nframes

number of animation frames. This only app

fps

frames per second. Default: 30 for tm_facets_animate and 2 for tm_facets_animate_slow.

play

how should the animation be played? One of "loop", "pingpong", and "once"

dpr

device pixel ratio. The ratio between the physical pixel density of a device and its logical pixel density.

...

passed on to tm_facets()

See also