Draw iso (contour) lines with labels

This function is a wrapper of tm_lines and tm_text aimed to draw isopleths.

tm_iso(
  col = NA,
  text = "level",
  size = 0.5,
  remove.overlap = TRUE,
  along.lines = TRUE,
  overwrite.lines = TRUE,
  bg.color = tmap_options()$bg.color,
  group = NA,
  ...
)

Arguments

col

line color. See tm_lines.

text

text to display.

size

text size (see tm_text)

remove.overlap

see tm_text

along.lines

see tm_text

overwrite.lines

see tm_text

bg.color

background color of the labels. Note: in tmap <= 3.2, the iso lines were cut to make space for labels. In tmap >= 3.3, this is changed: the iso lines remain unchanged, but the labels are printed with a background color by default.

group

name of the group to which this layer belongs in view mode. Each group can be selected or deselected in the layer control item. Set group = NULL to hide the layer in the layer control item. By default, it will be set to the name of the shape (specified in tm_shape).

...

arguments passed on to tm_lines or tm_text