tm_iso {tmap}	R Documentation

Draw iso (contour) lines with labels

Description

This function is a wrapper of tm_lines and tm_text aimed to draw isopleths.

Usage

tm_iso(
  col = NA,
  text = "level",
  size = 0.5,
  remove.overlap = TRUE,
  along.lines = TRUE,
  overwrite.lines = TRUE,
  group = NA,
  ...
)

Arguments

col
line color. See tm_lines.

text
text to display.

size
text size (see tm_text)

remove.overlap
see tm_text

along.lines
see tm_text

overwrite.lines
see tm_text

group
name of the group to which this layer belongs in view mode. Each group can be selected or deselected in the layer control item. Set group = NULL to hide the layer in the layer control item. By default, it will be set to the name of the shape (specified in tm_shape).

...
arguments passed on to tm_lines or tm_text

[Package tmap version 3.0 Index]