Minimap

Creates a minimap in view mode. See addMiniMap.

tm_minimap(server = NA, position = c("left", "bottom"), toggle = TRUE, ...)

Arguments

server

name of the provider or an URL (see tm_tiles). By default, it shows the same map as the basemap, and moreover, it will automatically change when the user switches basemaps. Note the latter does not happen when server is specified.

position

position of the scale bar Vector of two values, specifying the x and y coordinates. The first is either "left" or "right", the second either "top" or "bottom".

toggle

should the minimap have a button to minimise it? By default TRUE.

...

arguments passed on to addMiniMap.

See also