View mode options

View mode options. These options are specific to the view mode.

Usage

tm_view(
  use.WebGL,
  control.position,
  control.bases,
  control.overlays,
  set.bounds,
  set.view,
  set.zoom.limits,
  leaflet.options
)

Arguments

use.WebGL

use webGL layers with leafgl

control.position

position of the control attribute

control.bases

base layers

control.overlays

overlay layers

set.bounds

logical that determines whether maximum bounds are set, or a bounding box. Not applicable in plot mode. In view mode, this is passed on to setMaxBounds()

set.view

numeric vector that determines the view. Either a vector of three: lng, lat, and zoom, or a single value: zoom. See setView(). Only applicable if bbox is not specified

set.zoom.limits

numeric vector of two that set the minimum and maximum zoom levels (see tileOptions()).

leaflet.options

options passed on to leafletOptions()