Get or add format options

Format options are tmap options that are shape dependent. With tmap_format() the predefined formats can be retrieved. The values for a specific format can be retrieved with tmap_format(format), where format is the name of the format. The function tmap_format_add() is used to add a format.

Usage

tmap_format(format)

tmap_format_add(..., name)

Arguments

format

Name of the format. Run tmap_format() to see the choices.

...

Options from tm_layout() or tm_view(). Can also be a list of those options.

name

Name of the new format.

Value

The function tmap_format() returns the names of the available formats. When format is defined, it returns the option list corresponding the that format.

See also

Examples

# available formats
tmap_format()
#> available formats are: "World", "World_wide", "NLD", "NLD_wide" 

# create option list to be used as a new format
World_small = tmap_format("World")
World_small$scale = 2

# add format
tmap_format_add(World_small, name = "World_small")
#> format World_small succesfully added. Use this format with tm_format("World_small")

# observe that World_small is successfully added:
tmap_format()
#> available formats are: "World", "World_wide", "NLD", "NLD_wide", "World_small" 

data(World)

#qtm(World, fill="HPI", format="World_small")